Rename user/userSlice to auth/authSlice
