+++ src/features/loupe/FullSizeImage.jsx
@@ -2,9 +2,9 @@ import React, { useState, useRef, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useResizeObserver } from '../../app/utils';
 import { styled } from '../../theme/stitches.config';
-import { selectUserUsername, selectUserCurrentRoles } from '../user/userSlice';
+import { selectUserUsername, selectUserCurrentRoles } from '../auth/authSlice';
 import { selectIsDrawingBbox} from './loupeSlice';
-import { hasRole, WRITE_OBJECTS_ROLES, DELETE_IMAGES } from '../../auth/roles';
+import { hasRole, WRITE_OBJECTS_ROLES, DELETE_IMAGES } from '../auth/roles';
 import { selectWorkingImages, labelsValidated, markedEmpty } from '../review/reviewSlice';
 import { deleteImages } from '../images/imagesSlice';
 import { Image } from '../../components/Image';
